Core/Log: Fix some issues detected by static analysis.
* Either inefficient or wrong usage of string::find(). string::compare()
will
be faster if string::find's result is compared with 0, because it will
not scan the whole string.
If your intention is to check that there are no findings in the string,
you should compare with std::string::npos.

* C-style pointer casting detected. C++ offers four different kinds of
casts as replacements:
static_cast, const_cast, dynamic_cast and reinterpret_cast.